backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Brotli==1.1.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.0.1
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
fastapi==0.92.0
ffmpeg-python==0.2.0
fonttools==4.54.1
frozenlist==1.3.3
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.5.0
httpx==0.27.2
idna==3.4
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.2
multidict==6.0.4
mutagen==1.47.0
netifaces==0.11.0
numpy==2.0.2
openai==1.5.0
packaging==24.1
pillow==11.0.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==1.10.18
pyparsing==3.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.6
requests==2.28.2
six==1.16.0
sniffio==1.3.0
srt==3.5.3
starlette==0.25.0
tobiiglassesctrl==2.2.6
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
vosk==0.3.45
watchfiles==0.18.1
websockets==10.4
yarl==1.11.1
yt-dlp==2023.10.13
zipp==3.20.2
